More Logging (#26)
* Some minor error corrections
* Adds more logging in LogViewModel
** Logs appear in Infos & Help Screen now
** Add caller information (file name and method name):
** format: "filename::methodname: message"

+++ app/src/main/java/de/bikebean/app/ui/main/status/location/LocationStateRepository.java
@@ -63,8 +63,8 @@ class LocationStateRepository extends StateRepository {
     }
 
     void confirmWapp(Wapp wapp) {
-        confirmWapp(wapp.getCellTowers().getSmsId(), (double) wapp.getWifiAccessPoints().getSmsId());
-        confirmWapp(wapp.getWifiAccessPoints().getSmsId(), (double) wapp.getCellTowers().getSmsId());
+        confirmWapp(wapp.getCellTowers().getSmsId(), wapp.getWifiAccessPoints().getSmsId());
+        confirmWapp(wapp.getWifiAccessPoints().getSmsId(), wapp.getCellTowers().getSmsId());
     }
 
     private void confirmWapp(int smsId, double value) {
